ClassMapper v.1.1
added a oracle file that contains  a function (get_happiness()).

This function return the number of preferences satisfied, and also the number of unsatisfied.

The file Oracle1.py is dedicated to all analysis tools that can be used to analyze the performance of each algorithm.

Cheers! And welcome to version 1.1!